[FIX] mail, hr_holidays: correct "back on" indicator position
Before this commit, the "back on" indicator in the avatar card
resource popover would show up in the same line as the name.
This issue was introduced by [1] which changed the template of the
avatar card popover without adapting the ovverride in `hr_holidays`.
This commit fixes the issue by giving a `name` property to the avatar
card name element and adapting the override to it.

[1] https://github.com/odoo/odoo/pull/177445

closes odoo/odoo#239470

Signed-off-by: Giorgio Tocco (gito) <gito@odoo.com>